chore(package): move common files to prelude
affects: @fio/benchmarks, @fio/core, @fio/prelude
