+++ packages/benchmarks/package.json
@@ -32,6 +32,8 @@
     "@types/bluebird": "^3.5.28",
     "benchmark": "^2.1.4",
     "bluebird": "^3.7.1",
-    "fluture": "^11.0.3"
+    "fluture": "^11.0.3",
+    "@fio/core": "*",
+    "@fio/prelude": "*"
   }
 }
